@@ -1,23 +1,20 @@
 import 'package:dio/dio.dart';
-import 'package:flutter/foundation.dart';
 import 'api_response.dart';
 import 'api_route.dart';
 import 'decodable.dart';
 
 // ignore: one_member_abstracts
 abstract class BaseAPIClient {
-
   Future<ResponseWrapper<T>> request<T extends Decodable>({
-    @required APIRouteConfigurable route,
-    @required Create<T> create,
+    required APIRouteConfigurable route,
+    required Create<T> create,
     dynamic data,
   });
-
 }
-class APIClient implements BaseAPIClient {
 
+class APIClient implements BaseAPIClient {
   final BaseOptions options;
-  Dio instance;
+  Dio? instance;
 
   APIClient(this.options) {
     instance = Dio(options);
@@ -25,29 +22,25 @@ class APIClient implements BaseAPIClient {
 
   @override
   Future<ResponseWrapper<T>> request<T extends Decodable>({
-    @required APIRouteConfigurable route,
-    @required Create<T> create,
+    required APIRouteConfigurable route,
+    required Create<T> create,
     dynamic data,
   }) async {
-    
     final config = route.getConfig();
-    config.baseUrl = options.baseUrl;
-    config.data = data;
+    config?.baseUrl = options.baseUrl;
+    config?.data = data;
 
-    final response = await instance.fetch(config);
-    final responseData = response.data;
+    if (config == null) {
+      throw ErrorResponse(message: 'Failed to load request options.');
+    }
 
-    if (response.statusCode == 200) {
+    final response = await instance?.fetch(config);
+    final responseData = response?.data;
+
+    if (response?.statusCode == 200) {
       return ResponseWrapper.init(create: create, json: responseData);
     }
 
-    final errorResponse = ErrorResponse.fromJson(data) 
-      ?? ErrorResponse(
-        message: 'Request failed with status code: ${response.statusCode}'
-      );
-
-    throw errorResponse;
-
+    throw ErrorResponse.fromJson(data);
   }
-
 }
