@@ -1,15 +1,16 @@
+import 'package:app_structure/configs/constants.dart';
 import 'package:dio/dio.dart';
+import 'package:flutter/foundation.dart';
 
 import '../api_client.dart';
 import '../decodable.dart';
 
 class AuthToken implements Decodable<AuthToken> {
-
   String accessToken;
-  String refreshToken;
-  int expiredTime;
+  String? refreshToken;
+  int? expiredTime;
 
-  AuthToken({ this.accessToken, this.refreshToken, this.expiredTime });
+  AuthToken({required this.accessToken, this.refreshToken, this.expiredTime});
 
   @override
   AuthToken decode(dynamic data) {
@@ -18,46 +19,40 @@ class AuthToken implements Decodable<AuthToken> {
   }
 
   Future startRefreshToken() async {
-    await Future.delayed(Duration(seconds: 5));
+    await Future.delayed(const Duration(seconds: 5));
     // assign new access token
-    accessToken = 'eyadfj9803924jjdfkasjdfjsdf';
+    accessToken = '';
   }
 
   bool isExpired() {
     return true;
   }
-  
 }
 
 class AuthInterceptor extends InterceptorsWrapper {
-
   final APIClient client;
   AuthToken token;
 
-  AuthInterceptor(this.client, this.token); 
+  AuthInterceptor(this.client, this.token);
 
   @override
   Future onRequest(
-    RequestOptions options, 
-    RequestInterceptorHandler handler
-  ) async {
-
-    if (options.extra['no_auth'] ?? false) {
+      RequestOptions options, RequestInterceptorHandler handler) async {
+    final authorize = options.extra[RequestExtraKeys.authorize] ?? false;
+    if (!authorize) {
       return super.onRequest(options, handler);
     }
 
     if (token.isExpired()) {
-      client.instance.lock();
-      print('Lock request for refreshing token...');
+      client.instance?.lock();
+      debugPrint('Lock request for refreshing token...');
       await token.startRefreshToken();
-      client.instance.unlock();
-      print('Refresh token completed!');
+      client.instance?.unlock();
+      debugPrint('Refresh token completed!');
     }
 
     options.headers['Authorization'] = token.accessToken;
 
     return super.onRequest(options, handler);
-
   }
-
 }
