@@ -2,19 +2,19 @@ import 'package:flutter/material.dart';
 import '../../utils/app_colors.dart';
 
 class CommonAppBar extends StatelessWidget implements PreferredSizeWidget {
-
-  final String title;
-  final Widget titleWidget;
+  final String? title;
+  final Widget? titleWidget;
   final double elevation;
-  final TextStyle titleStyle;
-  final Color backgroundColor;
-  final Widget leftBarButtonItem;
-  final List<Widget> rightBarButtonItems;
+  final TextStyle? titleStyle;
+  final Color? backgroundColor;
+  final Widget? leftBarButtonItem;
+  final List<Widget>? rightBarButtonItems;
   final double titleSpacing;
   final bool automaticallyImplyLeading;
   final bool centerTitle;
 
-  CommonAppBar({
+  const CommonAppBar({
+    Key? key,
     this.title,
     this.titleStyle,
     this.titleWidget,
@@ -23,13 +23,12 @@ class CommonAppBar extends StatelessWidget implements PreferredSizeWidget {
     this.backgroundColor,
     this.leftBarButtonItem,
     this.rightBarButtonItems,
-    this.automaticallyImplyLeading = true, 
-    this.centerTitle,
-  });
+    this.automaticallyImplyLeading = true,
+    this.centerTitle = true,
+  }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
-
     return AppBar(
       title: _buildTitle(),
       elevation: elevation,
@@ -37,10 +36,10 @@ class CommonAppBar extends StatelessWidget implements PreferredSizeWidget {
       titleSpacing: titleSpacing,
       leading: _buildLeading(),
       automaticallyImplyLeading: automaticallyImplyLeading,
-      centerTitle: centerTitle ?? true,
+      centerTitle: centerTitle,
       actions: <Widget>[
         Container(
-          margin: EdgeInsets.fromLTRB(16, 0, 16, 0),
+          margin: const EdgeInsets.fromLTRB(16, 0, 16, 0),
           child: Row(
             mainAxisAlignment: MainAxisAlignment.spaceBetween,
             children: rightBarButtonItems ?? <Widget>[],
@@ -51,26 +50,21 @@ class CommonAppBar extends StatelessWidget implements PreferredSizeWidget {
   }
 
   @override
-  Size get preferredSize => Size.fromHeight(kToolbarHeight);
+  Size get preferredSize => const Size.fromHeight(kToolbarHeight);
 
   Widget _buildTitle() {
-
     final titleBarText = title ?? '';
-    final titleBarTextStyle = titleStyle 
-      ?? TextStyle(fontWeight: FontWeight.normal, fontSize: 17);
+    final titleBarTextStyle = titleStyle ??
+        const TextStyle(fontWeight: FontWeight.normal, fontSize: 17);
 
     return titleWidget ?? Text(titleBarText, style: titleBarTextStyle);
-
   }
 
   Widget _buildLeading() {
-
-    return (leftBarButtonItem != null) 
-      ? Container(
-        margin: EdgeInsets.fromLTRB(16, 0, 16, 0), 
-        child: leftBarButtonItem) 
-      : null;
-
+    return (leftBarButtonItem != null)
+        ? Container(
+            margin: const EdgeInsets.fromLTRB(16, 0, 16, 0),
+            child: leftBarButtonItem)
+        : Container();
   }
-
 }
