@@ -10,31 +10,29 @@ import 'cubit/app_cubit.dart';
 import 'cubit/app_state.dart';
 
 class App extends StatefulWidget {
-
+  const App({Key? key}) : super(key: key);
   @override
   _AppState createState() => _AppState();
 }
 
 // ignore: prefer_mixin
 class _AppState extends State<App> with WidgetsBindingObserver {
-
   final Routes routes = Routes();
 
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance?.addObserver(this);
   }
 
   @override
   void dispose() {
-    WidgetsBinding.instance.removeObserver(this);
+    WidgetsBinding.instance?.removeObserver(this);
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-
     return MultiBlocProvider(
       providers: [
         BlocProvider(
@@ -44,17 +42,16 @@ class _AppState extends State<App> with WidgetsBindingObserver {
       child: GestureDetector(
         onTap: () {
           //Hide keyboard when tap outside
-          WidgetsBinding.instance.focusManager.primaryFocus?.unfocus();
+          WidgetsBinding.instance?.focusManager.primaryFocus?.unfocus();
         },
         child: BlocBuilder<AppCubit, AppState>(
           builder: (context, state) {
             return MaterialApp(
-              locale: 
-                (state is AppLanguageFetchLocaleCompleted) 
-                ? state.locale 
-                : Locale('en'),
+              locale: (state is AppLanguageFetchLocaleCompleted)
+                  ? state.locale
+                  : const Locale('en'),
               supportedLocales: AppLocalizations.supportedLocales,
-              localizationsDelegates: [
+              localizationsDelegates: const [
                 AppLocalizations.delegate,
                 GlobalMaterialLocalizations.delegate,
                 GlobalWidgetsLocalizations.delegate,
@@ -67,6 +64,6 @@ class _AppState extends State<App> with WidgetsBindingObserver {
           },
         ),
       ),
-    ); 
+    );
   }
 }
