@@ -1,11 +1,10 @@
 import 'package:bloc/bloc.dart';
+import 'package:flutter/foundation.dart';
 
 class AppBlocObserver extends BlocObserver {
-  
   @override
   void onTransition(Bloc bloc, Transition transition) {
     super.onTransition(bloc, transition);
-    print('onTransition $transition');
+    debugPrint('onTransition $transition');
   }
-
 }
