@@ -1,8 +1,10 @@
 class Keystores {
-
   static const String accessToken = 'AccessTokenKeystores';
   static const String userRole = 'UserRoleKeystore';
   static const String userInfo = 'UserInfoKeystore';
   static const String userLocale = 'UserLocale';
+}
 
+class RequestExtraKeys {
+  static const String authorize = 'Authorize';
 }
