@@ -5,8 +5,6 @@ import 'configs/flavor_config.dart';
 
 void main() {
   FlavorConfig(
-    flavor: Flavor.dev, 
-    values: FlavorValues(baseUrl: 'https://abc.com/api')
-  );
-  runApp(App());
+      flavor: Flavor.dev, values: FlavorValues(baseUrl: 'https://abc.com/api'));
+  runApp(const App());
 }
