@@ -1,10 +1,13 @@
 import 'package:flutter/material.dart';
 
 class EventDetailsScreen extends StatelessWidget {
+  const EventDetailsScreen({Key? key}) : super(key: key);
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(title: Text('Event Details'),),
+      appBar: AppBar(
+        title: const Text('Event Details'),
+      ),
       body: Container(),
     );
   }
